\name{inline-package}
\alias{inline-package}
\alias{inline}
\docType{package}
\title{\packageTitle{inline}}
\description{\packageDescription{inline}}
\seealso{\code{\link{cfunction}}, \code{\link{cxxfunction}}}
\author{\packageAuthor{inline}}
\section{Maintainer}{\packageMaintainer{inline}}
\keyword{package}

