\name{internals}
\alias{internals}

\alias{setCMethod,character,list,list-method}
\alias{setCMethod,character,character,character-method}

\title{ Internals }

\description{
   Aliases required for 'R CMD check' but those noone will ever search for.
   Here to prevent the mess of the index. The corresponding items have help
   aliases without method signature!
}

\keyword{internal}
