\name{inline-package}
\alias{inline-package}
\alias{inline}
\docType{package}
\title{
Inline C, C++, Fortran function calls from R
}
\description{
Functionality to dynamically define R functions and S4 methods with in-lined C, 
C++ or Fortran code supporting \code{\link{.C}} and \code{\link{.Call}}
calling conventions.
}
\details{
\tabular{ll}{
Package: \tab inline\cr
Type: \tab Package\cr
Version: \tab 0.3.8\cr
Date: \tab 2010-12-07\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
}
\seealso{
	\code{\link{cfunction}}, \code{\link{cxxfunction}}	
}
\author{
Oleg Sklyar, Duncan Murdoch, Mike Smith, Dirk Eddelbuettel, Romain Francois

Maintainer: Oleg Sklyar <oleg.sklyar@googlemail.com>
}
\keyword{ package }

